2.6.187: more typescript migration, login form fixes, added mac-filter in interface.ssid.radius.authentication
