+++ public/locales/pt/translation.json
@@ -523,6 +523,7 @@
 	},
 	"login": {
 		"access_policy": "Política de Acesso",
+		"change_password_error": "Senha rejeitada, talvez seja uma senha antiga",
 		"change_password_explanation": "Digite e confirme sua nova senha",
 		"change_your_password": "Mudar senha",
 		"confirm_new_password": "confirme a nova senha",
